Fix command for dry-run in README
In our newest versions of snakemake, the dryrun command as written will fail because of that -r. Even in our oldest snakemake module (7.25.0), this flag is deprecated (-r or --reason prints the reason for each executed rule -- deprecated because it's always true now). Removing the -r resolves errors and works regardless of which module is used.